Add files via upload
I implemented the function "eval()" for to get the value entered by user and to convert the expression of the string in an expression valid. This value its shown on the display by a method that takes the private attribute and places the new value. Too i has places two new methods, one for calc the expression and another for looks into array the last number entered by user and places it on display.